fix check style
