+++ test/plugin/scenarios/grizzly-2.3.x-4.x-scenario/src/main/java/test/apache/skywalking/apm/testcase/grizzly/controller/HealCheckHandler.java
@@ -1,0 +1,34 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package test.apache.skywalking.apm.testcase.grizzly.controller;
+
+import org.glassfish.grizzly.http.server.HttpHandler;
+import org.glassfish.grizzly.http.server.Request;
+import org.glassfish.grizzly.http.server.Response;
+
+public class HealCheckHandler extends HttpHandler {
+
+    @Override
+    public void service(Request request, Response response) throws Exception {
+        String hello = "hello";
+        response.setContentType("text/plain");
+        response.setContentLength(hello.length());
+        response.getWriter().write(hello);
+    }
+}
